separate the plugin to export the user's render function from email templates into a new file
